Deploy Calculate 2024/01/07

+++ Interactor/SuperAdmin/TerminateTenantInteractor.cs
@@ -1,9 +1,11 @@
 ﻿using AWSSDK.Common;
 using AWSSDK.Constants;
+using AWSSDK.Dto;
 using AWSSDK.Interfaces;
 using Domain.SuperAdminModels.Notification;
 using Domain.SuperAdminModels.Tenant;
 using Interactor.Realtime;
+using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Configuration;
 using UseCase.SuperAdmin.TerminateTenant;
 
@@ -17,6 +19,7 @@ namespace Interactor.SuperAdmin
         private readonly ITenantRepository _tenantRepositoryRunTask;
         private readonly INotificationRepository _notificationRepositoryRunTask;
         private readonly IConfiguration _configuration;
+        private readonly IMemoryCache _memoryCache;
 
         public TerminateTenantInteractor(
             ITenantRepository tenantRepository,
@@ -24,7 +27,8 @@ namespace Interactor.SuperAdmin
             INotificationRepository notificationRepository,
             ITenantRepository tenantRepositoryRunTask,
             INotificationRepository notificationRepositoryRunTask,
-            IConfiguration configuration
+            IConfiguration configuration,
+             IMemoryCache memoryCache
             )
         {
             _awsSdkService = awsSdkService;
@@ -33,6 +37,7 @@ namespace Interactor.SuperAdmin
             _tenantRepositoryRunTask = tenantRepositoryRunTask;
             _notificationRepositoryRunTask = notificationRepositoryRunTask;
             _configuration = configuration;
+            _memoryCache = memoryCache;
         }
         public TerminateTenantOutputData Handle(TerminateTenantInputData inputData)
         {
@@ -41,7 +46,7 @@ namespace Interactor.SuperAdmin
                 IWebSocketService _webSocketService;
                 _webSocketService = (IWebSocketService)inputData.WebSocketService;
 
-                string pathFileDumpTerminate = _configuration["PathFileDumpTerminate"];
+                string pathFileDumpTerminate = _configuration["PathFileDumpTerminate"] ?? string.Empty;
 
                 if (string.IsNullOrEmpty(pathFileDumpTerminate))
                 {
@@ -59,47 +64,75 @@ namespace Interactor.SuperAdmin
                     return new TerminateTenantOutputData(false, TerminateTenantStatus.TenantDoesNotExist);
                 }
 
-                var listTenantDb = RDSAction.GetListDatabase(tenant.EndPointDb).Result;
-                // Check valid delete tennatDb
-                if (listTenantDb == null || listTenantDb.Count() == 0 || !listTenantDb.Contains(tenant.Db))
+                if (tenant.Status == ConfigConstant.StatusTenantDictionary()["terminating"])
                 {
-                    return new TerminateTenantOutputData(false, TerminateTenantStatus.TenantDbDoesNotExistInRDS);
+                    return new TerminateTenantOutputData(false, TerminateTenantStatus.TenantIsTerminating);
                 }
 
-                _tenantRepository.UpdateStatusTenant(inputData.TenantId, ConfigConstant.StatusTenantDictionary()["terminating"]);
+                if (tenant.Status == ConfigConstant.StatusTenantDictionary()["terminating"])
+                {
+                    return new TerminateTenantOutputData(false, TerminateTenantStatus.TenantIsTerminating);
+                }
 
+                if ((tenant.StatusTenant == ConfigConstant.StatusTenantPending || tenant.StatusTenant == ConfigConstant.StatusTenantStopping || tenant.StatusTenant == ConfigConstant.StatusTenantFailded) && inputData.Type == 1)
+                {
+                    return new TerminateTenantOutputData(false, TerminateTenantStatus.TenantIsNotAvailableToSortTerminate);
+                }
+                _tenantRepository.UpdateStatusTenant(inputData.TenantId, ConfigConstant.StatusTenantDictionary()["terminating"]);
+                var messenge = tenant.EndSubDomain + $"がシャットダウン中です。";
+                _webSocketService.SendMessageAsync(FunctionCodes.SuperAdmin, new NotificationModel(tenant.TenantId, ConfigConstant.StatusNotiSuccess, ConfigConstant.StatusSuttingDown, messenge));
                 CancellationTokenSource cts = new CancellationTokenSource();
                 _ = Task.Run(() =>
                 {
                     try
                     {
+                        bool skipFinalSnapshot = false;
+                        var listTenantDb = RDSAction.GetListDatabase(tenant.EndPointDb, tenant.UserConnect, tenant.PasswordConnect).Result;
                         Console.WriteLine($"Start Terminate tenant: {tenant.RdsIdentifier}");
 
+                        if (tenant.StatusTenant == ConfigConstant.StatusTenantPending || tenant.StatusTenant == ConfigConstant.StatusTenantStopping) // terminate tenant  creating, updating, restoring
+                        {
+                            skipFinalSnapshot = true;
+                            var tenantInfo = _memoryCache.Get<TenantCacheMemory>(tenant.SubDomain) ?? new TenantCacheMemory();
+
+                            // Cancel task run
+                            tenantInfo.CancelToken.Cancel();
+
+                            // Delete tmp RDS
+                            if (!string.IsNullOrEmpty(tenantInfo.TmpRdsIdentifier))
+                            {
+                                if (RDSAction.CheckRDSInstanceExists(tenantInfo.TmpRdsIdentifier).Result)
+                                {
+                                    var deleteTmpRDSAction = RDSAction.DeleteRDSInstanceAsync(tenantInfo.TmpRdsIdentifier, skipFinalSnapshot).Result;
+                                }
+                            }
+                        }
+
                         // Backup tenant
-                        if (inputData.Type == 1) // sort terminate
+                        if (inputData.Type == 1)
                         {
                             // Create folder backup S3
-                            var backupFolder = @$"bk-{tenant.EndSubDomain}";
-                            _awsSdkService.CreateFolderBackupAsync(ConfigConstant.DestinationBucketName, tenant.EndSubDomain, ConfigConstant.DestinationBucketName, backupFolder).Wait();
-                            
+                            var backupFolderName = @$"bk-{tenant.EndSubDomain}";
+                            _awsSdkService.CreateFolderBackupAsync(ConfigConstant.DestinationBucketName, tenant.EndSubDomain, ConfigConstant.RestoreBucketName, backupFolderName).Wait();
+
                             // Dump DB backup
-                            var pathFileDump = @$"{pathFileDumpTerminate}{tenant.Db}.sql"; // path save file sql dump
-                           PostgresSqlAction.PostgreSqlDump(pathFileDump, tenant.EndPointDb, ConfigConstant.PgPostDefault, tenant.Db, "postgres", "Emr!23456789").Wait();
-                            
+                            var pathFileDump = @$"{pathFileDumpTerminate}{tenant.Db}.sql"; // Path save file sql dump
+                            PostgresSqlAction.PostgreSqlDump(pathFileDump, tenant.EndPointDb, ConfigConstant.PgPostDefault, tenant.Db, ConfigConstant.PgUserDefault, ConfigConstant.PgPasswordDefault).Wait();
+
                             // check valid file sql dump
                             if (!System.IO.File.Exists(pathFileDump))
                             {
-                                throw new Exception("File sql dump doesn't exist");
+                                throw new Exception("sqldump 存在しません。");
                             }
 
                             long length = new System.IO.FileInfo(pathFileDump).Length;
                             if (length <= 0)
                             {
-                                throw new Exception("Invalid file sql dump");
+                                throw new Exception("Sqldump が無効です");
                             }
-                            
+
                             // Upload file sql dump to folder backup S3
-                            _awsSdkService.UploadFileAsync(ConfigConstant.RestoreBucketName, $@"{backupFolder}/{tenant.Db}", pathFileDump).Wait();
+                            _awsSdkService.UploadFileAsync(ConfigConstant.RestoreBucketName, $@"{backupFolderName}/{tenant.Db}", pathFileDump).Wait();
                         }
 
                         bool deleteRDSAction = false;
@@ -107,7 +140,11 @@ namespace Interactor.SuperAdmin
                         // Connect RDS delete TenantDb
                         if (listTenantDb.Count > 1)
                         {
-                            if (_awsSdkService.DeleteTenantDb(tenant.EndPointDb, tenant.Db))
+                            if (listTenantDb.Contains(tenant.Db))
+                            {
+                                deleteRDSAction = _awsSdkService.DeleteTenantDb(tenant.EndPointDb, tenant.Db, tenant.UserConnect, tenant.PasswordConnect);
+                            }
+                            else
                             {
                                 deleteRDSAction = true;
                             }
@@ -116,11 +153,34 @@ namespace Interactor.SuperAdmin
                         // Deleted RDS
                         else
                         {
-                            deleteRDSAction = RDSAction.DeleteRDSInstanceAsync(tenant.RdsIdentifier).Result;
+                            if (RDSAction.CheckRDSInstanceExists(tenant.RdsIdentifier).Result)
+                            {
+                                // Check RDS being used by another tenant
+                                if (_tenantRepositoryRunTask.GetByRdsId(tenant.TenantId, tenant.RdsIdentifier).Count() == 1)
+                                {
+                                    deleteRDSAction = RDSAction.DeleteRDSInstanceAsync(tenant.RdsIdentifier, skipFinalSnapshot).Result;
+                                }
+                                else
+                                {
+                                    deleteRDSAction = true;
+                                }
+                            }
+                            else
+                            {
+                                deleteRDSAction = true;
+                            }
                         }
 
                         // Delete DNS
-                        var deleteDNSAction = Route53Action.DeleteTenantDomain(tenant.SubDomain).Result;
+                        bool deleteDNSAction = false;
+                        if (Route53Action.CheckSubdomainExistence(tenant.SubDomain).Result) // Check exist DNS
+                        {
+                            deleteDNSAction = Route53Action.DeleteTenantDomain(tenant.SubDomain).Result;
+                        }
+                        else
+                        {
+                            deleteDNSAction = true;
+                        }
 
                         // Delete item cname in cloud front
                         var deleteItemCnameAction = CloudFrontAction.RemoveItemCnameAsync(tenant.SubDomain).Result;
@@ -136,9 +196,18 @@ namespace Interactor.SuperAdmin
                             {
                                 _tenantRepositoryRunTask.TerminateTenant(inputData.TenantId, ConfigConstant.StatusTenantDictionary()["terminated"]);
                                 // Notification  terminating success
-                                var messenge = tenant.EndSubDomain + $"is teminate successfully. ";
+                                var messenge = tenant.EndSubDomain + $"の終了が完了しました。";
                                 var notification = _notificationRepositoryRunTask.CreateNotification(ConfigConstant.StatusNotiSuccess, messenge);
+
+                                // Add info tenant for notification
+                                notification.SetTenantId(tenant.TenantId);
+                                notification.SetStatusTenant(ConfigConstant.StatusTenantTeminated);
+
                                 _webSocketService.SendMessageAsync(FunctionCodes.SuperAdmin, notification);
+
+                                // Delete cache memory
+                                _memoryCache.Remove(tenant.SubDomain);
+
                                 cts.Cancel();
                                 return;
                             }
@@ -146,11 +215,20 @@ namespace Interactor.SuperAdmin
                     }
                     catch (Exception ex)
                     {
-                        _tenantRepositoryRunTask.TerminateTenant(inputData.TenantId, ConfigConstant.StatusTenantDictionary()["terminate-failed"]);
+                        _tenantRepositoryRunTask.UpdateStatusTenant(inputData.TenantId, ConfigConstant.StatusTenantDictionary()["terminate-failed"]);
                         // Notification  terminating failed
-                        var messenge = $"{tenant.EndSubDomain} is teminate failed. Error: {ex.Message}.";
+                        var messenge = $"{tenant.EndSubDomain} の終了に失敗しました。エラー: {ex.Message}.";
                         var notification = _notificationRepositoryRunTask.CreateNotification(ConfigConstant.StatusNotifailure, messenge);
+
+                        // Add info tenant for notification
+                        notification.SetTenantId(tenant.TenantId);
+                        notification.SetStatusTenant(ConfigConstant.StatusTenantFailded);
+
                         _webSocketService.SendMessageAsync(FunctionCodes.SuperAdmin, notification);
+
+                        // Delete cache memory
+                        _memoryCache.Remove(tenant.SubDomain);
+
                         cts.Cancel();
                         return;
                     }
